+++ cocos/2d/CCTMXXMLParser.cpp
@@ -271,7 +271,7 @@ void TMXMapInfo::startElement(void* /*ctx*/, const char *name, const char **atts
         }
 
 
-        float hexSideLength = attributeDict["hexsidelength"].asFloat();
+        int hexSideLength = attributeDict["hexsidelength"].asInt();
         tmxMapInfo->setHexSideLength(hexSideLength);
 
         Size s;
@@ -356,7 +356,7 @@ void TMXMapInfo::startElement(void* /*ctx*/, const char *name, const char **atts
             TMXLayerInfo* layer = tmxMapInfo->getLayers().back();
             Size layerSize = layer->_layerSize;
             uint32_t gid = static_cast<uint32_t>(attributeDict["gid"].asUnsignedInt());
-            int tilesAmount = layerSize.width*layerSize.height;
+            int tilesAmount = (int)(layerSize.width*layerSize.height);
             
             if (_xmlTileIndex < tilesAmount)
             {
@@ -417,7 +417,6 @@ void TMXMapInfo::startElement(void* /*ctx*/, const char *name, const char **atts
         TMXTilesetInfo* tileset = tmxMapInfo->getTilesets().back();
         
         float tileOffsetX = attributeDict["x"].asFloat();
-        
         float tileOffsetY = attributeDict["y"].asFloat();
         
         tileset->_tileOffset = Vec2(tileOffsetX, tileOffsetY);
@@ -452,7 +451,7 @@ void TMXMapInfo::startElement(void* /*ctx*/, const char *name, const char **atts
             
             TMXLayerInfo* layer = tmxMapInfo->getLayers().back();
             Size layerSize = layer->_layerSize;
-            int tilesAmount = layerSize.width*layerSize.height;
+            int tilesAmount = (int)(layerSize.width*layerSize.height);
 
             uint32_t *tiles = (uint32_t*) malloc(tilesAmount*sizeof(uint32_t));
             // set all value to 0
@@ -512,8 +511,8 @@ void TMXMapInfo::startElement(void* /*ctx*/, const char *name, const char **atts
         dict["x"] = Value(p.x);
         dict["y"] = Value(p.y);
         
-        int width = attributeDict["width"].asInt();
-        int height = attributeDict["height"].asInt();
+        float width = attributeDict["width"].asFloat();
+        float height = attributeDict["height"].asFloat();
         Size s(width, height);
         s = CC_SIZE_PIXELS_TO_POINTS(s);
         dict["width"] = Value(s.width);
@@ -694,7 +693,7 @@ void TMXMapInfo::endElement(void* /*ctx*/, const char *name)
                 unsigned char *deflated = nullptr;
                 Size s = layer->_layerSize;
                 // int sizeHint = s.width * s.height * sizeof(uint32_t);
-                ssize_t sizeHint = s.width * s.height * sizeof(unsigned int);
+                ssize_t sizeHint = static_cast<ssize_t>(s.width * s.height * sizeof(unsigned int));
                 
                 ssize_t CC_UNUSED inflatedLen = ZipUtils::inflateMemoryWithHint(buffer, len, &deflated, sizeHint);
                 CCASSERT(inflatedLen == sizeHint, "inflatedLen should be equal to sizeHint!");
